chore: move translateQueryData func from token utils to helper utils
