+++ src/router/newOrder.ts
@@ -2,7 +2,8 @@ import * as _ from 'lodash'
 import { getPrice } from '../request/marketMaker'
 import { PriceApiResult } from '../request/marketMaker/interface'
 import { getFormatedSignedOrder } from '../utils/order'
-import { getSupportedTokens, translateQueryData } from '../utils/token'
+import { getSupportedTokens } from '../utils/token'
+import { translateQueryData } from '../utils/helper'
 import { updaterStack } from '../utils/intervalUpdater'
 import { QueryInterface } from './interface'
 import { checkParams } from '../validations'
